adding projects page
