+++ src/pages/index.tsx
@@ -3,7 +3,7 @@ import Layout from '../components/Layout';
 import Header from '../components/Header';
 import Landing from '../sections/Landing';
 import About from '../sections/About';
-// import Projects from '../sections/Projects';
+import Projects from '../sections/Projects';
 // import Writing from '../sections/Writing';
 // import Footer from '../components/Footer';
 
@@ -12,8 +12,8 @@ const IndexPage = () => (
     <Header />
     <Landing />
     <About />
-    {/*<Projects />
-    <Writing />
+    <Projects />
+    {/*<Writing />
     <Footer /> */}
   </Layout>
 );
