+++ src/sections/Landing.tsx
@@ -46,9 +46,9 @@ const LandingPage = () => {
       </Heading>
 
       <Flex alignItems="center" justifyContent="center" flexWrap="wrap">
-        {socialLinks.map(({ id, fontAwesomeIcon, ...rest }) => (
-          <Box mx={3} fontSize={[5, 6, 6]} key={id}>
-            <SocialLink icon={fontAwesomeIcon} {...rest} />
+        {socialLinks.map((sl) => (
+          <Box mx={3} fontSize={[5, 6, 6]} key={sl.name}>
+            <SocialLink {...sl} />
           </Box>
         ))}
       </Flex>
